Simplified and documented the tuple database API. Changed TTL to an int value, measured in seconds.

+++ sesamestream-api/src/main/java/edu/rpi/twc/sesamestream/Subscription.java
@@ -30,5 +30,5 @@ public interface Subscription {
      * @param ttl the new time-to-live in milliseconds, or 0 for inifinite time-to-live
      * @return whether the subscription was successfully renewed
      */
-    boolean renew(long ttl);
+    boolean renew(int ttl);
 }
